+++ Changelog.md
@@ -2,3 +2,6 @@
 
 ## Changes
 - Search box should be usable in more modded chests now
+
+## Fixes
+- Dimmed and highlighted slots are aligned incorrectly in shulker boxes